refactor(tonk-display): extract target-agnostic tonk-template crate
The template planner (segment parsing, the Binding/PlanNode/
BindingPlan types, path arithmetic, and per-row string
substitution) was already pure std-only Rust living inside the
tonk-display wasm cdylib. Move it into a new target-agnostic
tonk-template crate so a headless renderer can share the exact same
planner: same template in, same plan out, by construction.

tonk-display now depends on tonk-template and re-exports it, keeping
its crate::template::* paths resolving; it retains only the
browser-only half (DOM walk to collect bindings, DOM apply). Also
scaffolds tonk-render (the future headless renderer) with a tl
parser spike confirming tl exposes text+element children with
DOM-matching child indices, including whitespace text nodes.
